requirements.txt
fastapi
uvicorn
openai
torch>=1.10
transformers
pillow
python-dotenv
